benchmark/pyproject.toml
[tool.poetry]
name = "imagefilter-benchmark"
version = "1.0.0"
description = "Бенчмарк тесты для измерения производительности фильтров изображений ImageFilter"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "imagefilter_benchmark"}]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^10.0.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
benchmark = "imagefilter_benchmark.benchmark:main"
generate-images = "imagefilter_benchmark.generate_test_images:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py37']

[tool.ruff]
line-length = 100
target-version = "py37"

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

